Extracting enum, renaming type
