+++ src/features/attachments/upload/uploadAttachmentSagas.ts
@@ -12,12 +12,13 @@ import { httpPost } from 'src/utils/network/networking';
 import { isAxiosError } from 'src/utils/network/sharedNetworking';
 import { fileUploadUrl } from 'src/utils/urls/appUrlHelper';
 import { customEncodeURI } from 'src/utils/urls/urlHelper';
-import { BackendValidationSeverity, getValidationMessage } from 'src/utils/validation/backendValidation';
+import { getValidationMessage } from 'src/utils/validation/backendValidation';
+import { BackendValidationSeverity } from 'src/utils/validation/backendValidationSeverity';
 import type { IAttachment } from 'src/features/attachments';
 import type { IUploadAttachmentAction } from 'src/features/attachments/upload/uploadAttachmentActions';
 import type { IUseLanguage } from 'src/hooks/useLanguage';
 import type { IRuntimeState } from 'src/types';
-import type { IComponentValidations, IValidationIssue } from 'src/utils/validation/types';
+import type { BackendValidationIssue, IComponentValidations } from 'src/utils/validation/types';
 
 export function* uploadAttachmentSaga({
   payload: { file, attachmentType, tmpAttachmentId, componentId, dataModelBindings, index },
@@ -90,7 +91,7 @@ export function* uploadAttachmentSaga({
     let validations: IComponentValidations;
 
     if (backendFeatures?.jsonObjectInDataResponse && isAxiosError(err) && err.response?.data?.result) {
-      const validationIssues: IValidationIssue[] = err.response.data.result;
+      const validationIssues: BackendValidationIssue[] = err.response.data.result;
 
       validations = {
         simpleBinding: {
