+++ src/utils/validation/backendValidation.ts
@@ -1,17 +1,9 @@
+import { BackendValidationSeverity } from 'src/utils/validation/backendValidationSeverity';
 import { buildValidationObject, unmappedError } from 'src/utils/validation/validationHelpers';
 import { validationTexts } from 'src/utils/validation/validationTexts';
 import type { IUseLanguage } from 'src/hooks/useLanguage';
 import type { LayoutPages } from 'src/utils/layout/LayoutPages';
-import type { IValidationIssue, IValidationObject, ValidationSeverity } from 'src/utils/validation/types';
-
-export enum BackendValidationSeverity {
-  Unspecified = 0,
-  Error = 1,
-  Warning = 2,
-  Informational = 3,
-  Fixed = 4,
-  Success = 5,
-}
+import type { BackendValidationIssue, IValidationObject, ValidationSeverity } from 'src/utils/validation/types';
 
 /**
  * We need to map the severity we get from backend into the format used when storing in redux.
@@ -35,7 +27,7 @@ export enum ValidationIssueSources {
  * Some validations performed by the backend are also performed by the frontend.
  * We need to ignore these to prevent duplicate errors.
  */
-export function shouldExcludeValidationIssue(issue: IValidationIssue): boolean {
+export function shouldExcludeValidationIssue(issue: BackendValidationIssue): boolean {
   /*
    * Legacy required validation detection
    * Remove this condition in v4, assuming that a minimum backend version is required.
@@ -68,7 +60,11 @@ export function shouldExcludeValidationIssue(issue: IValidationIssue): boolean {
 /**
  * Gets standard validation messages for backend validation issues.
  */
-export function getValidationMessage(issue: IValidationIssue, langTools: IUseLanguage, params?: string[]): string {
+export function getValidationMessage(
+  issue: BackendValidationIssue,
+  langTools: IUseLanguage,
+  params?: string[],
+): string {
   const { langAsString } = langTools;
   if (issue.customTextKey) {
     return langAsString(issue.customTextKey, params);
@@ -98,7 +94,7 @@ export function getValidationMessage(issue: IValidationIssue, langTools: IUseLan
  * Maps validation issues from the backend into the intermediate format used by the frontend.
  */
 export function mapValidationIssues(
-  issues: IValidationIssue[],
+  issues: BackendValidationIssue[],
   resolvedNodes: LayoutPages,
   langTools: IUseLanguage,
 ): IValidationObject[] {
