+++ src/utils/validation/backendValidationSeverity.ts
@@ -1,0 +1,8 @@
+export enum BackendValidationSeverity {
+  Unspecified = 0,
+  Error = 1,
+  Warning = 2,
+  Informational = 3,
+  Fixed = 4,
+  Success = 5,
+}
